DROOLS-3978 Move codegen unit tests
