+++ drools/kie-dmn/kogito-dmn-rest-utils/src/main/java/org/kie/dmn/kogito/rest/quarkus/DMNDecisionInfo.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2019 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kie.dmn.kogito.rest.quarkus;
+
+import java.io.Serializable;
+
+import javax.json.bind.annotation.JsonbProperty;
+
+import org.kie.dmn.api.core.ast.DecisionNode;
+
+public class DMNDecisionInfo implements Serializable {
+
+    private String id;
+    private String name;
+
+    public DMNDecisionInfo() {
+        // Intentionally blank.
+    }
+
+    public static DMNDecisionInfo of(DecisionNode decisionNode) {
+        DMNDecisionInfo res = new DMNDecisionInfo();
+        res.setName(decisionNode.getName());
+        res.setId(decisionNode.getId());
+        return res;
+    }
+
+    @JsonbProperty("decision-id")
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    @JsonbProperty("decision-name")
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
