--- drools/kie-dmn/kogito-dmn-rest-utils/src/main/java/org/kie/dmn/submarine/rest/quarkus/DMNUnaryTestsInfo.java
@@ -1,49 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.dmn.submarine.rest.quarkus;
-
-import java.io.Serializable;
-
-import javax.json.bind.annotation.JsonbProperty;
-
-public class DMNUnaryTestsInfo implements Serializable {
-
-    private String text;
-    private String expressionLanguage;
-
-    public DMNUnaryTestsInfo() {
-        // Intentionally blank.
-    }
-
-    @JsonbProperty("unarytests-text")
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-
-    @JsonbProperty("unarytests-expressionLanguage")
-    public String getExpressionLanguage() {
-        return expressionLanguage;
-    }
-
-    public void setExpressionLanguage(String expressionLanguage) {
-        this.expressionLanguage = expressionLanguage;
-    }
-}
